reorganize src/ to reflect kernel vs frontend vs agent boundaries
Move shell.ts, input-handler.ts, output-parser.ts into src/shell/ —
these are the interactive terminal frontend, not core kernel.

Move token-budget.ts into src/agent/ — only used by agent-loop,
hardcoded overhead values specific to internal agent's prompt structure.

Move builtin-extensions.ts into src/extensions/index.ts — manifest
belongs with the extensions it loads.
