+++ src/extensions/index.ts
@@ -5,7 +5,7 @@
  * They receive unscoped contexts (not reloadable) and can be individually
  * disabled via the `disabledBuiltins` setting in ~/.agent-sh/settings.json.
  */
-import type { ExtensionContext } from "./types.js";
+import type { ExtensionContext } from "../types.js";
 
 type ActivateFn = (ctx: ExtensionContext) => void;
 
@@ -13,14 +13,14 @@ export const BUILTIN_EXTENSIONS: Array<{
   name: string;
   load: () => Promise<ActivateFn>;
 }> = [
-  { name: "agent-backend",    load: () => import("./extensions/agent-backend.js").then(m => m.default) },
-  { name: "tui-renderer",     load: () => import("./extensions/tui-renderer.js").then(m => m.default) },
-  { name: "slash-commands",    load: () => import("./extensions/slash-commands.js").then(m => m.default) },
-  { name: "file-autocomplete", load: () => import("./extensions/file-autocomplete.js").then(m => m.default) },
-  { name: "shell-recall",     load: () => import("./extensions/shell-recall.js").then(m => m.default) },
-  { name: "command-suggest",   load: () => import("./extensions/command-suggest.js").then(m => m.default) },
-  { name: "terminal-buffer",   load: () => import("./extensions/terminal-buffer.js").then(m => m.default) },
-  { name: "overlay-agent",     load: () => import("./extensions/overlay-agent.js").then(m => m.default) },
+  { name: "agent-backend",    load: () => import("./agent-backend.js").then(m => m.default) },
+  { name: "tui-renderer",     load: () => import("./tui-renderer.js").then(m => m.default) },
+  { name: "slash-commands",    load: () => import("./slash-commands.js").then(m => m.default) },
+  { name: "file-autocomplete", load: () => import("./file-autocomplete.js").then(m => m.default) },
+  { name: "shell-recall",     load: () => import("./shell-recall.js").then(m => m.default) },
+  { name: "command-suggest",   load: () => import("./command-suggest.js").then(m => m.default) },
+  { name: "terminal-buffer",   load: () => import("./terminal-buffer.js").then(m => m.default) },
+  { name: "overlay-agent",     load: () => import("./overlay-agent.js").then(m => m.default) },
 ];
 
 /**
